backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
redis==5.0.1
gunicorn==21.2.0
flask-caching==2.1.0
Werkzeug==2.3.7 